minor changes and comments
